change: positions and size based on Payload#DPI

+++ RecorteHorarioSemanalIFROJipa/PayloadClassTimeTable.cs
@@ -17,8 +17,8 @@
 
 public class PayloadClassTimeTable
 {
-    public int X { get; set; }
-    public int? Y { get; set; }
-    public int? Width { get; set; }
-    public int? Height { get; set; }
+    public double X { get; set; }
+    public double? Y { get; set; }
+    public double? Width { get; set; }
+    public double? Height { get; set; }
 }
